+++ modules/functions_manager/tests/test_private/test_private_2/private/findprivate.m
@@ -1,0 +1,6 @@
+function r = findprivate()
+ % test_private_2 private function 2.
+ disp('You found the private function test_private_2.')
+ r = 5 + toto();
+end
+%=============================================================================